Squash Ever's work and a few of my commits.
Commit of the MAD ccpi analysis with the new branches

Changes with that is possible send the mracho to the grid

Some modifications where I added counters for the number of the events that pass each cut, I add the kVtx cut that is used for the cases where the fiducial volume cut of the branche is not the same that for CCPionInc. There are some errors, I'm checking how to fix it

Errors fixed, the code its working now

Adding the function to plot the ratios, and some small changes to run the analysis with to validate the new tuplas

It is used to check the number of subruns with a exit status != 0.

CVUniverse with the old Ehad definition

I made some modifications to try to add the particle response systematics. In the end I removed the changes

I made changes to run the code with different playlists

I made some changes to get the ratio plots for the different playlists

I made some modifications for the ratios plotter

I made some changes to just run the playlists ME1A-ME1D

I made some changes to to obtain the old Ehad definition

Changes to run the new root files for the reduced playlists

Changes to plot the new results

Re-adding MAT-compatibility, one file at a time.

Manual merge of the MAT-compatible stuff into Ever's current work.

Leaves Ever's work untouched except for a few places where file location changed,
and I removed some event counting stuff from makeCrossSectionMCInputs. Ever is
more than welcome to add that stuff back in.

Couple updates to setup scripts.