The end of the muon study, at least.
The answer is that the mad branch vars get you px and py in the
detector coordinates, which is NOT what you want. Nor does it
propagate systematics.

Indeed you want things in beam coordinates, which you then unfold
to neutrino direction.

What is the future of this branch? Probably just let it hang.