Large Systematics Update, Add Pmu SD Constraint (#21)
Added/updated the following systematics:

* Node Cut Efficiency
* Particle Response
* Update Muon Systematics
* Update Genie Systematics
* Michel Efficiency
* Diffractive Pion

Target mass systematic not added because I don't think it applies.

There are also some non-systematics-related updates that snuck in as well, add pmu signal definition constraint, being the main one.

--------------------------------------------------------------------------------------

* Implementation of the ehad variable

* Modifications to aplay the Aaron's Node eff Systematics

* Implementation  of Node eff Aaron's Cut

* Remove extra line.

* Clean up some unwanted/unused code.

* Remove unwanted file.

* Addition of the Aaron's Systematics, The Node eff cut systematics should be checked.

* Update of tuple-utils/SubmitAuditAllPlaylistsToGrid.sh

* Changes to try to fix the Node Eff Systematics, Not working yet.

* Adition of the new systemats plots

* Reformat, clean up, add override.

* Remove old ifndef.

* Just reformat.

* Important update to make node cut eff systematic work.

Must call get weight after cuts function because node cut eff syst weight
requires pion candidate.

Second important update: universe::m_is_truth is static, so we need to
set it separately for all univs when we do reco loop and again when
we do truth loop.

Neither of these are good solutions, but I'll clean up later.

And speaking of clean up, I also did some reformatting, as usual.

* Fix the node cut efficiency systematic (#14)

There were two problems with this systematic. It's a weird syst. It's vertical only (affecting only getweight), which isn't weird in itself, but it does require a pion candidate, which means that for this universe specifically, get weight must be called after the cuts have been made. That's new and it means we can't call getweight within ccpievent -- which is annoying.

The second weird thing that we've never been bitten by before is that it asks whether this IsTruth. But what i didn't realize is that IsTruth is static. So we need to manually set the truth status of each universe before we do the reco and then truth loops. Very awkward. Will want to clean that up some day.

Otherwise, just some misc clean ups.

But the resulting node cut eff for ME1L looks good (non-zero, non-huge, non-flat).

* Reformat, clean up, add override.

* Remove old ifndef.

* Just reformat.

* Important update to make node cut eff systematic work.

Must call get weight after cuts function because node cut eff syst weight
requires pion candidate.

Second important update: universe::m_is_truth is static, so we need to
set it separately for all univs when we do reco loop and again when
we do truth loop.

Neither of these are good solutions, but I'll clean up later.

And speaking of clean up, I also did some reformatting, as usual.

* Fix particle response systematic.

Needed to include the parent functions from RecoilEnergyFunctions, to
swap GetNonCalRecoilEnergy for GetTrackRecoilEnergy, and to set the
track candidate vector with SetNonCalIndices.

* Add some new muon systematics (#15)

This merge doesn't actually do anything -- it only changes some comments and removes unused stuff. I'm completing this PR in hopes that it will make the next PR easier.

Co-authored-by: Everardo Granados Vazquez <granados@minervagpvm07.fnal.gov>
Co-authored-by: egranadosvazquez <granadosve2012@licifug.ugto.mx>

* Update and add new genie systematics (#16)

Use the latest and greatest GetGenie systematics functions.

* includes/Systematics.h

* Adition of Angle Systematics

* Modifications to add the GENIE MaRES and NormCCRes

* Modifications to add the GENIE MaRES and NormCCRes

* Cleanup.

* Cleanup.

* Cleanup.

Co-authored-by: Everardo Granados Vazquez <granados@minervagpvm07.fnal.gov>
Co-authored-by: egranadosvazquez <granadosve2012@licifug.ugto.mx>

* Add Pmu Signal Definition Constraint (#18)

Add event selection cut and signal definition constraint for 1.5 < pmu < 20 GeV.

Also uncomment michel systematics (but they're not turned on yet, because need to add the weight to `CVU::GetWeight`.

* Remove no-longer-needed ifndefs.

* includes/Systematics.h

* Adition of Angle Systematics

* Modifications to add the GENIE MaRES and NormCCRes

* Modifications to add the GENIE MaRES and NormCCRes

* Changes to the pmu binning variable to observe that the new cut is apply

* Changes to add the new cut for pmu

* Changes to add the new cut for pmu

* Addition of Michel Systematics

* Changes for the pmu signal definition

* Modifications to add the kpmu cut

* Cleanup.

* kAllCuts must come last in enum declaration.

Co-authored-by: Everardo Granados Vazquez <granados@minervagpvm07.fnal.gov>
Co-authored-by: egranadosvazquez <granadosve2012@licifug.ugto.mx>

* Add Standard Michel Efficiency Weight to GetWeight

The systematic was already added to the systematics list, but we need to add it's weight to CVUniverse::GetWeight to turn it on.

* Implement Diffractive Systematics (#19)

Couple comments on this implementation: (1) probably don't need to consider nuclear target material diffractive universes. (2) this universe implementation belongs in MAT itself, not here. Leave it for now.

* aplicacion of the Diffractive weight and adition of some functions to calculate the weight

* aplicacion of the Diffractive weight and adition of some functions to calculate the weight

* aplication of the Diffractive systematics

* Diffractive systematics files

* Diffractive systematics files

* Adition of CohDiffractiveSystematics.cxx

* Adition of Diffractive systematics plots

* Just undo an unimportant commit that will confuse the PR.

Co-authored-by: Everardo Granados Vazquez <granados@minervagpvm06.fnal.gov>
Co-authored-by: Everardo Granados Vazquez <granados@minervagpvm07.fnal.gov>
Co-authored-by: egranadosvazquez <granadosve2012@licifug.ugto.mx>
Co-authored-by: egranadosvazquez <50630075+egranadosvazquez@users.noreply.github.com>