Ever's Michel changes to MAD.
These changes should be relative to the previous version.

Strategy for merging this version into current version will be
to first see if git can do it gracefully. If not, I'll look at the
diff between this version and the previous to see if I can
do it manually without too much trouble.