GetDistMichel Bugfix.
GetDistMichel had been accessing fit distances by cluster ID. Instead
it should be accessing them by vertex. I believe this changed when
we switched over to MAD.

The impact of the bug is complicated. Evidently, we only completely
failed to get a fit distance a few % of the time. Otherwise, when
we succeeded in getting a sensible fit distance, I think it might
have been wrong, maybe 33-66% of the time? Curious to see the
impact of this fix on the results. Bizarre.