Copy of Ever's branch with a single commit for easier rebasing and comparison to master.
In this branche I'm adding the Aaron's 12 binnig, I made the changes to sent jos to the grid. I'm waiting for the results of the things tha I sent to the grid to check if this is fixed

Modiffications to sent jobs to the grid

I'm adding the Deuterium GeniePi Tune, Probably it could solve the discrepancy bettwen Ben's and Aaron analysis

Checking if the weights are applied correctly, I observed that probably the v4 was not applied corectly

I change the function that get the Low q2 weight for the same function that Aaron used. I add to xsec/plotCrossSectionFromFile.C the code to make the rations and comparizons with Aaron's thesis results

Making some changes to the plotting ratio function

Update Grid Tarball Handling

The way the grid handles tarballs has changed.

Update crossSectionDataFromFile.C

removing the changes

Update plotCrossSectionFromFile.C

Changing the values of the Aaron's cross section for the publication values

I change the Y axis scale and the cross section values form Aaron's thesis to the Aaron's paper

Changes to plot q2 GeV  in logaritmic scale

Autoformat binning, constants, and macroutil.

Autoformat.