New branch squashing all of Ever's commits from the CutsOrderAaronSigDef branch into one for easier rebasing and comparison to master.
In this branche I'm adding the Aaron's 12 binnig, I made the changes to sent jos to the grid. I'm waiting for the results of the things tha I sent to the grid to check if this is fixed

Modiffications to sent jobs to the grid

I'm adding the Deuterium GeniePi Tune, Probably it could solve the discrepancy bettwen Ben's and Aaron analysis

Checking if the weights are applied correctly, I observed that probably the v4 was not applied corectly

In this updates I made the nessesary modifications to have the same fata selection and signal definition that Aaron. I checked the events that are signal and I compared with the Aarons signal definition for 1000 events of the tupla /pnfs/minerva/persistent/users/granados/MADtuplas/merged/20211115/mc/ME1A/MasterAnaDev_mc_AnaTuple_run00110000_Playlist.root. With thsi update I'm obtaining the same signal events and the same weight

The weight fot the q2 < 0 is 1

Applying the Aaron's fid volume. I just made this change and something weid happens. When I try to run xsec/crossSectionDataFromFile.C and it doesn't converge when try to find the fit for the background. So we need to understand what's happening there. I changed the ntuplas for the p3 sample

I'm changing the dedfinition of the Ehad function, In this moment I'm using the function GetRecoilEnergy() that comes from PlotUtils/RecoilEnergyFunctions.h

I modiffy the q2 function as Aaron has it

I'm checking the how are applied the cuts and I'm ordering these in the same way that are in the Aaron Analysis

I'm changing the order to print the breackdown weights

Revert unwanted changes. Remove Q2GeV stuff.

Clean up.

Revert Q2 GeV changes mostly.