Addressing a memory leak in the trackless Michels.
Initializing class data members at the declaration has always
been sketchy to me, and I think it caused a problem in this case.

Specifically, the passable_matchtype vector of ints I think was
causing this problem. And related to the way the Michel CTOR was
being called? IDK. Changing the Michel CTOR call to use new keyword
has solved the problem for now. But long term, I either need to
understand why it failed or re-write it in what I know to be a safe

way.