Closure Test Tpi, Fix Closure Flux Norm (#51)
Hard code in the ability to to Tpi or Pmu closure. A todo would be to pass the variable to the program rather than hard code it.

Also, importantly, caught a flux norm bug. I think this had been a problem, but we had convinced ourselves to ignore it before? Anyway, I think we understand it now.

* Making closure test for Tpi

* Fixing the ratio flux normalization

---------

Co-authored-by: egranadosvazquez <granadosve2012@licifug.ugto.mx>