Reduce binning of w fit variable.
I think the fit had been failing due to empty bins.

The reduced binning size may still be a problem for
some universes, but it's working for the CV.

I had also been failing to set the POT scale in
macroutil. IDK what impact that had or if it ever
mattered. Fixed now in any case.