PassesCuts Now Returns a Struct (#68)
We now need PassesCuts to return a fourth object, `bool passes_all_cuts_except_w`, which is too many.

So: pack them all into a new `struct PassesCutsInfo`.

The analysis pipeline is updated accordingly, as is `runSidebands`, but other scripts will have to be updated later.

* Introduce simple PassesCutsInfo struct.

Also need passes cuts to return passes_all_except_w. That puts
the number of returned variables up to 4, which is too many, so
put them all into this struct.

Attempt to fully implement the new function in the pipeline
as well as in runSidebands.