fix debug mode.
Debug mode had not been working. So hadn't noticed a typo in a michel assert:
endpoint::GetQualityMichels should find a vtx matched michel anywhere except
for the vertex.

In FillCounters, during the cut loop, break out if we fail a cut instead of continue.
Is that correct? I gotta think about it.