ECalRecoil No-Pi Fix (#81)
In GetCalRecoilEnergyNoPi_Corrected, which gets called by Ehad, we apply Aaron's correction to non-tracked, not-coherent events. I had been erroneously only checking pions with idx > 0 for coherence. I don't see why events with idx = 0 shouldn't also be included. Hadron indexes can be 0, right? It's my michel branches that have a special case of the 0th index. But not hadron indexes. This has an impact on results. Bigger than I would have expected. It should just very few events in or out of the W cut. I should check it at greater statistics later.

Also: update a couple checks of no-pion-candidates in the Erecoil-adjacent functions. At the moment, it captures edge cases. If you forget to set pion candidates when Wexp is called this will come up. Or if you call the Wexp cut out of order or if you don't call michel cuts at all. This will all come up. Gonna definitely have to revisit this when we merge in trackless::michels, though.