All of Ever's commits combined into 1
Initial commit of making MAT michels work.

Add comments.

Improved comments.

I made some changes to start to work with MAT trackless events

I made some changes to plot the backgorund and data selection, I added some cuts to the selection, currently it has a purity of 65%

I'm omminting changes related to use the trackless pions in tracked pion analysis.

In this commit is included the different binnings that I have been trying to use. I made the changes to always use the kinetic energy for the primary pions, not secondary pions

In this branch I'm making a correction to Ehad calculation

I make some small changes to remove the data points in the histograms

I change the addition to the ehad that I was doing, currently it only add the rest mass of the pion

I just trurned off the MaRES weight