Move GetPlaylistFile to MacroUtil. Add test playlist option. No physics changes. (#93)
Remove dependence on ccpion_common. Move the playlist access file to MacroUtil.h.

Add the "run_test_playlist" argument to makeMCXSecInputs and crossSectionData for a small playlist. Add this arg to the Process python script.

Eventually make this macroutil switch for all scripts, and remove ccpion_common forever.

* Working commit.

* Give up on this static map idea.

* Abandon the idea of a MakeMacroUtil function.

* Undo unwanted commits.

* PrintMacroConfig bugfix.

* ProcessCCPiMacro needs the new do_test_playlist option.

Which will always be false for this script.

* Separate Loop function changes from the i/o macroutil changes.

Only the latter are for this branch. Incoming branch for the former.

* Bugfix.

* Make the file access changes for crossSectionData as well.