Another disasterous working commit.
Some progress with the loop function. Moved the expensive reco
and cuts stuff into an event::process function.

Now, struggling with how to not break the ability to loop
over variables and call GetValue on it.

Still big todo: the actual work of getting pion variables
and comparing untracked michels with tracked michels.